Support component versioning
By default components must have unique names base only on class, this is good because things with the same name should not do different things
and as such fail fast. However when you want to version components and run them concurrently its useful that if they exist in a versioned package
then they don't clash, e.g. a.b.v1.C and a.b.v2.C should be OK but a.b.C and a.d.C should not be.